requirements-py310.txt
pillow==11.1.0
pytesseract==0.3.10
opencv_python==4.11.0.86
ultralytics==8.3.87
transformers==4.49.0
pdf2image==1.17.0
doctr==1.9.0
python-doctr==0.11.0
beautifulsoup4==4.12.2
matplotlib==3.10.1
streamlit==1.43.2
